requirements.txt
streamlit
torch
torchvision
transformers
pandas
plotly
pillow
numpy
